requirements.txt
django == 4.1.5
djangorestframework == 3.14.0
markdown == 3.4.1
django-filter == 22.1